Validate each subtitle block
A subtitle block is validated by checking for an index and a timecode
line at the beginning of the block. The index is a simple integer. The
time code line consists of two time stamps in this format:

HH:MM:SS,mmm --> HH:MM:SS,mmm

where HH is hours, MM is minutes, SS is seconds and mmm are
milliseconds.