Rewrite subtitle class
The subtitle class has been substantially modified. The index of the
subtitle is publicly accessible and can be modified directly. The start
time and stop time are read-only and the dialog is entirely private.

The subtitle is instantiated using a text block, an array of strings
read directly from an SRT file. The empty? method will answer whether
the subtitle is valid or discardable. The subtitle can be written
directly to a file or stdout in proper SRT format.