Avoid creating invalid SRT files after delay
If delay is too negative, the timecode will turn negative. This is
tested for and prevented.

Similarly, if the delay is too positive, the last timecode will extend
past 23:59:59,999. What should happen here is not very clear. We do know
that the hour must be two digits. If we simply convert days into hours,
we could go on for more than four days. If we don't, the maximum for the
stop time is the one day mark. For now, this is the mark used to limit
the delay on the positive side.

It is also ensured that the read-in subtitles are rejected if timecodes
exceed past one day.